Search behaviour review and code cleanup
If you delete a search query nice and slow, it will show no books when the search field is empty. However, if you delete the search query fast, it will display the results rendered on the first "backspace" keystroke or "delete" keystroke.  Ctrl+a and delete removes the query in the search field and no books are shown on the search page.